refactor(style): re-order project structure
