+++ src/components/colliders/BaseCollider.vue
@@ -2,7 +2,7 @@
 import { inject, nextTick, onUnmounted, type ShallowRef, shallowRef, watch } from 'vue'
 
 import { useRapierContext } from '../../composables'
-import { createCollider } from '../../utils/collider'
+import { createCollider } from '../../core/collider'
 import type { ColliderProps, CreateColliderReturnType, ExposedCollider, RigidBodyContext } from '../../types'
 
 const props = withDefaults(defineProps<Partial<ColliderProps>>(), {
