+++ src/core/index.ts
@@ -1,5 +1,3 @@
-export * from '../components'
-export * from '../composables'
-export * from '../constants'
-export * from '../types'
-export * from '../utils'
+export * from './collider'
+export * from './injection-keys'
+export * from './rigid-body'
