+++ src/core/rigid-body.ts
