+++ src/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './collider'
-export * from './rigid-body'
